updates st_nn_ont_proj.R to fix {units} bug
updated line 33 to `udunits_from_proj[st_crs(x)$units]` from standard `st_crs(x)$units`. I think this was leading to an error from the units::as_units.call() function:
```
 Error: cannot convert ft into us
Did you try to supply a value in a context where a bare expression was expected? 
```
The fix in {sf} for a similar issue as using this manner to get to the units. This approach returns:
```
$`us-ft`
1 [US_survey_foot]
```
instead of the former method (st_crs(x)$units) which returns:
```
[1] "us-ft"
```
This all came of issue when calling nngeo::st_nn in spatial join.